chore(ui-ux): organize address page components
